+++ src/pages/settings/index.tsx
@@ -1,14 +1,12 @@
+import Profile from '@/components/Forms/Profile'
+import ProfilePageMenu from '@/components/Menu/ProfilePageMenu'
 import Footer from '@/components/WebLayout/Footer'
 import Header from '@/components/WebLayout/Header'
+import { SubmenuType } from '@projectType/componentTypes'
 import { useState } from 'react'
 
-const Profile = () => {
-  const [slug, setSlug] = useState<string>('')
-
-  type SubmenuType = {
-    name: string
-    slug: string
-  }[]
+const ProfilePage = () => {
+  const [slug, setSlug] = useState<string>('profil')
 
   const submenu: SubmenuType = [
     { name: 'Profil', slug: 'profil' },
@@ -25,58 +23,9 @@ const Profile = () => {
           <h1 className="text-6xl font-semibold text-white">Nastavení</h1>
         </section>
         <section className="w-full mt-20 shadow-sl flex">
-          <ul className="w-52 h-fit grid grid-cols-1">
-            {submenu.map((item) => (
-              <li
-                key={item.slug}
-                className={
-                  item.slug === slug
-                    ? 'h-9 pl-5 mr-3 py-1 text-xl cursor-pointer bg-streetlaw-500 text-white relative'
-                    : 'h-9 pl-5 py-1 text-xl cursor-pointer hover:bg-streetlaw-500 hover:text-white relative'
-                }
-                onClick={() => setSlug(item.slug)}
-              >
-                {item.slug === slug && (
-                  <div
-                    className="w-0 h-0 
-  border-t-[18px] border-t-transparent
-  border-l-[12px] border-l-streetlaw-500
-  border-b-[18px] border-b-transparent
-  absolute -right-3 top-0"
-                  ></div>
-                )}
-                {item.name}
-              </li>
-            ))}
-          </ul>
+          <ProfilePageMenu setSlug={setSlug} slug={slug} submenu={submenu} />
           <div className="border-l-[1px] w-[1085px] border-black">
-            <form className="grid grid-cols-5 items-center pl-9 pr-16 mt-7 gap-7">
-              <label htmlFor="email" className="profile-label">
-                Email:
-              </label>
-              <input id="email" type="email" className="profile-input" />
-              <label htmlFor="name" className="profile-label">
-                Jméno:
-              </label>
-              <input id="name" type="text" className="profile-input" />
-              <label htmlFor="surname" className="profile-label">
-                Příjmení:
-              </label>
-              <input id="surname" type="text" className="profile-input" />
-              <label htmlFor="about" className="profile-label place-self-start">
-                Popis:
-              </label>
-              <textarea
-                id="about"
-                cols={64}
-                rows={10}
-                className="col-span-4 resize-none overflow-y-scroll outline-none py-1 px-3 text-2xl bg-gray-100"
-              ></textarea>
-              <label htmlFor="photo" className="profile-label">
-                Fotografie:
-              </label>
-              <input type="file" />
-            </form>
+            <Profile />
           </div>
         </section>
       </div>
@@ -85,4 +34,4 @@ const Profile = () => {
   )
 }
 
-export default Profile
+export default ProfilePage
